ENH: merging back 4051:4426 changed. svn merge -r 4051:4425 https://gdcm.svn.sourceforge.net/svnroot/gdcm/trunk + manual 4426

+++ Utilities/VTK/vtkImageColorViewer.cxx
@@ -794,7 +794,7 @@ double vtkImageColorViewer::GetOverlayVisibility()
 
 void vtkImageColorViewer::SetOverlayVisibility(double vis) 
 {
-  this->OverlayImageActor->SetVisibility(vis);
+  this->OverlayImageActor->SetVisibility((int)vis);
 }
 
 //----------------------------------------------------------------------------
